services/diarization/pyproject.toml
[project]
name = "diarization-service"
version = "0.1.0"
description = "Speaker diarization service for Verbatim Studio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "httpx>=0.26.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "pyannote.audio>=3.1.0",
    "speechbrain>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
